requirements.txt
Django>=4.0.0,<4.1.0
# psycopg2-binary==2.9.1
# psycopg2>=2.8
djangorestframework==3.13.0
celery==5.2.3
django-cors-headers==3.6.0
django-celery-results==2.2.0
requests==2.22.0
martor==1.6.8
django-decouple==2.1
django-allauth==0.49.0
django-rest-auth==0.9.5
dj-rest-auth==2.2.3
SQLAlchemy==1.4.31
redis==4.1.3
django-filter==21.1
Markdown==3.3.4
django-import-export==2.7.0
Pillow==9.0.1